Refactored frontend into separate components
