+++ frontend/src/components/LoadingSpinner.tsx
@@ -1,0 +1,6 @@
+export const LoadingSpinner = () => (
+    <div className="loading-container">
+      <div className="spinner"></div>
+      <p>Loading...</p>
+    </div>
+  );
