+++ frontend/src/components/Modal.tsx
@@ -1,0 +1,21 @@
+import { ReactNode } from 'react';
+
+interface ModalProps {
+  title: string;
+  onClose: () => void;
+  children: ReactNode;
+}
+
+export const Modal = ({ title, onClose, children }: ModalProps) => (
+  <div className="modal-overlay" onClick={onClose}>
+    <div className="modal-content" onClick={e => e.stopPropagation()}>
+      <div className="modal-header">
+        <h2>{title}</h2>
+        <button onClick={onClose} className="close-button">&times;</button>
+      </div>
+      <div className="modal-body">
+        {children}
+      </div>
+    </div>
+  </div>
+);
